// sneaker-tracker/apps/web-nextjs/types/sneaker.ts
/*
 *   Copyright (c) 2025 
 *   All rights reserved.
 */

export interface SneakerRelease {
  id: string;
  name?: string;
  title?: string;
  productName?: string;
  date?: string;
  price?: string | number;
  mileage?: number;
  brand?: string;
  sku?: string;
  status?: string;
  locations?: string[];
}